+++ app/src/main/java/io/treehouses/remote/Fragments/SystemFragment.java
@@ -26,6 +26,9 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import io.treehouses.remote.Constants;
+import io.treehouses.remote.Fragments.DialogFragments.ChPasswordDialogFragment;
+import io.treehouses.remote.Fragments.DialogFragments.ContainerDialogFragment;
+import io.treehouses.remote.Fragments.DialogFragments.RenameDialogFragment;
 import io.treehouses.remote.Network.BluetoothChatService;
 import io.treehouses.remote.R;
 import io.treehouses.remote.bases.BaseFragment;
