+++ app/src/main/java/io/treehouses/remote/Network/BluetoothChatService.java
@@ -38,7 +38,7 @@ import java.io.Serializable;
 import java.util.UUID;
 
 import io.treehouses.remote.Fragments.HomeFragment;
-import io.treehouses.remote.Fragments.RPIDialogFragment;
+import io.treehouses.remote.Fragments.DialogFragments.RPIDialogFragment;
 import io.treehouses.remote.Constants;
 
 /**
